+++ lib/ui/emulator/emulator_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:nesd/ui/emulator/display.dart';
-import 'package:nesd/ui/emulator/touch_controls.dart';
+import 'package:nesd/ui/emulator/input/touch/touch_controls.dart';
 import 'package:nesd/ui/settings/settings.dart';
 
 class EmulatorWidget extends ConsumerWidget {
@@ -11,10 +11,10 @@ class EmulatorWidget extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final settings = ref.watch(settingsControllerProvider);
 
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
+    return Stack(
       children: [
-        const Expanded(child: DisplayWidget()),
+        const DisplayWidget(),
+        if (settings.showTouchControls) const TouchControls(),
       ],
     );
   }