+++ lib/ui/settings/controls/controls_settings.dart
@@ -7,6 +7,7 @@ import 'package:nesd/ui/common/settings_tile.dart';
 import 'package:nesd/ui/emulator/input/action/all_actions.dart';
 import 'package:nesd/ui/emulator/input/intents.dart';
 import 'package:nesd/ui/settings/controls/binding_tile.dart';
+import 'package:nesd/ui/settings/controls/show_touch_controls_switch.dart';
 import 'package:nesd/ui/settings/settings.dart';
 import 'package:nesd/ui/settings/settings_tab.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
@@ -22,6 +23,7 @@ class ControlsSettings extends StatelessWidget {
       index: 3,
       child: Column(
         children: [
+          const ShowTouchControlsSwitch(),
           const ProfileSelectionHeader(),
           Expanded(
             child: ListView(