+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Android support
 - Performance improvements
 - Support for small and narrow screens
+- Touch screen controls
 
 ## [0.2.0] - 2024-07-23